refactor tutor issues
