+++ project/src/mocks/map-cities.ts
