+++ project/src/store/reducer.ts
@@ -1,19 +1,19 @@
 import {createReducer} from '@reduxjs/toolkit';
 import {changeCity, filterCity} from './action';
 import {OFFERS} from '../mocks/offers';
-import {MAP_CITIES} from '../mocks/map';
+import {MAP_CITIES} from '../mocks/map-cities';
 import {reviews} from '../mocks/reviews';
 import {DEFAULT_CITY} from '../const';
 import {getFilterOffers, getFilterCity} from './utils';
 
-const intialState = {
+const initialState = {
   city: DEFAULT_CITY,
   offers: getFilterOffers(OFFERS, DEFAULT_CITY),
   reviews: reviews,
   mapCity: getFilterCity(MAP_CITIES, DEFAULT_CITY),
 };
 
-const reducer = createReducer(intialState, (builder) => {
+const reducer = createReducer(initialState, (builder) => {
   builder
     .addCase(changeCity, (state, action) => {
